move acct mapping logic to helper file for ct and agentless
